Rename _constants.py to _enums.py
